==> 02_Analyses/02c_ER_P.Rmd <==
---
title: "ER Analyses"
author: "JMH"
date: "11/19/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r echo= FALSE}
library(tidyverse)
library(here)
library(mgcViz)
library(MuMIn)
library(mgcv)
library(forecast)
library(gridExtra)
library(tidymv)
library(ggpubr)
```

# Load & munge data
```{r}
load(file.path(here::here("04_SavedImages"), "02c_ERanalysesPonly_rdat"))
```


```{r message=FALSE, warning=FALSE}
met <- readr::read_csv(file = file.path(here::here("01_Data"),"01_MundgedMetabDat.csv")) %>% 
  mutate(stream = fct_relevel(stream, c("st11U", "st18", "st9", "st6")),
         treatment = fct_relevel(treatment, c("ambient", "nitrogen", "phosphorus")),
         Yf = as.factor(Yf),
         streamTreat = as.factor(streamTreat)) %>% 
  filter(treatment == "ambient" | treatment == "phosphorus") %>% 
  droplevels()



```



# ER Plots
Big differences among treatments. Normalish
```{r echo=FALSE}
ggplot(met, aes(y = lER, x = doy, color = treatment)) +
  geom_point()+
  stat_smooth() +
  facet_wrap(vars(stream))
```


```{r}
ggplot(met, aes(sample = lER)) +stat_qq() + stat_qq_line() 
```


```{r}
ggplot(met, aes(sample = lER, color = stream)) +stat_qq() + stat_qq_line() +
  facet_grid(.~stream)
```



same as N, best model chosen for all didn't fit
```{r echo= FALSE}
BestERCor <- corARMA(p = 1,q = 1, form = ~ doy | streamTreat)
```


## Model sel

```{r echo=FALSE, include = T}
# g.modsel_t <- uGamm(lER ~ treatment * invKT.C.StMean  + s(LightPerDay.lcg),
#                 correlation = BestERCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# g.modsel_s <- uGamm(lER ~ treatment * stream  + s(LightPerDay.lcg),
#                 correlation = BestERCor, REML = FALSE, data = met,
#                 control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# 
# g.Dredge_t <-  dredge(global.model = g.modsel_t, evaluate = TRUE, rank = AICc)
# g.Dredge_s <-  dredge(global.model = g.modsel_s, evaluate = TRUE, rank = AICc)
```



```{r echo=FALSE}
g.DredgeSumF <- merge(g.Dredge_t, g.Dredge_s)
```

Model selection for all models
```{r echo=FALSE}
subset(g.DredgeSumF, delta <= 2)
```

```{r}
write_csv(g.DredgeSumF,
          file.path(here::here("06_AICtables"), "02a_ER_P.csv"))
```

```{r}
g.er.BEST <- gamm(lER ~ treatment * stream + s(LightPerDay.lcg),
                correlation = BestERCor, REML = TRUE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
```


Stream Temperature models; AIC difference = -31.2
```{r echo=FALSE}
# 151.5 v. 186.0
subset(g.Dredge_t, delta <= 2)
```


Best with stream temp
```{r}
g.er.BEST_t <- gamm(lER ~ treatment + invKT.C.StMean  + s(LightPerDay.lcg),
                correlation = BestERCor, REML = TRUE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
```



### Best models
#### Overall


```{r}
summary(g.er.BEST$gam)
```

```{r}
plot(g.er.BEST$gam, pages = 1)
```


Concurvity issues with Qres and DOYres
```{r}
round(concurvity(g.er.BEST$gam, full = FALSE)$estimate,3)
```

```{r}
g.er.BEST.gv <- getViz(g.er.BEST$gam)
check(g.er.BEST.gv)
```


One last check on the autocorrelation. Just foucsing on 3. Some remaining in S9A, S18A
```{r echo= FALSE}
g.er.best.acf <- met %>% 
  mutate(g.erRes = resid(g.er.BEST$lme, type = "normalized")) %>% 
  select(streamTreat, g.erRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.erRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.er.best.acf$acf, ncol = 3)
```



#### Best with temp

```{r}
summary(g.er.BEST_t$gam)
```

```{r}
plot(g.er.BEST_t$gam, pages = 1)
```


Concurvity issues with Qres and DOYres
```{r}
round(concurvity(g.er.BEST_t$gam, full = FALSE)$estimate,3)
```

```{r}
g.er.BEST_t.gv <- getViz(g.er.BEST_t$gam)
check(g.er.BEST_t.gv)
```



```{r echo=FALSE}
metBestTest <- met %>% 
  mutate(erres1 = residuals(g.er.BEST$gam, type = "response"),
         erPred1 = predict(g.er.BEST$gam, type = "response"),
         erres2 = residuals(g.er.BEST_t$gam, type = "response"),
         erPred2 = predict(g.er.BEST_t$gam, type = "response"))
```


 Check residuals
Still some signal with doy. Also early dates with S18-P giving problems
Not much of a difference between 1 and 2.
```{r echo=FALSE}
ggplot(metBestTest, aes(y = erres1, x = doy, color = stream, shape = treatment)) +
  geom_point() +
  facet_grid(stream ~ treatment) 
```

```{r echo=FALSE}
ggplot(metBestTest, aes(y = lER, x = erPred1, color = stream, shape = treatment)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1)
```


#### stream temp


```{r}
summary(g.er.BEST_t$gam)
```

Looking good
```{r}
round(concurvity(g.er.BEST_t$gam, full = FALSE)$estimate,3)
```

Looks pretty good, issue with one stream
```{r}
g.er.BEST_t.gv <- getViz(g.er.BEST_t$gam)
check(g.er.BEST_t.gv)
```

Still some autocorrelation remaining in S6A, S9A, S9P, S18A
```{r echo= FALSE}
g.BEST_t.acf <- met %>% 
  mutate(g.erRes = resid(g.er.BEST_t$lme, type = "normalized")) %>% 
  select(streamTreat, g.erRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.erRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.BEST_t.acf$acf, ncol = 3)
```

```{r echo=FALSE}
ggplot(metBestTest, aes(y = erres2, x = doy, color = stream, shape = treatment)) +
  geom_point() +
  facet_grid(stream ~ treatment) 
```

Not doing a great job with S18
```{r echo=FALSE}
ggplot(metBestTest, aes(y = lER, x = erPred2, color = stream, shape = treatment)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1)
```



### save results
```{r echo = FALSE, include = FALSE}
# saveRDS(g.er.BEST,
#         file = file.path(here::here("05_SavedModels"), "02a_ER_P_Best.rds"))
# saveRDS(g.er.BEST_t,
#         file = file.path(here::here("05_SavedModels"), "02a_ER_P_BestTemp.rds"))
```


```{r}
# save.image(file.path(here::here("04_SavedImages"), "02c_ERanalysesPonly_rdat"))
```
