==> 02_Analyses/04c_NEPanalyses_P.Rmd <==
---
title: "NEP Analyses"
author: "JMH"
date: "11/19/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r echo= FALSE}
library(tidyverse)
library(here)
library(mgcViz)
library(MuMIn)
library(mgcv)
library(forecast)
library(gridExtra)
```

# Load & munge data
```{r}
load(file.path(here::here("04_SavedImages"), "04c_NEPanalysesP_rdat"))
```


DOY < 183 already removed
```{r message=FALSE, warning=FALSE}
met <- readr::read_csv(file = file.path(here::here("01_Data"),"01_MundgedMetabDat.csv")) %>% 
  mutate(stream = fct_relevel(stream, c("st11U", "st18", "st9", "st6")),
         treatment = fct_relevel(treatment, c("ambient", "nitrogen", "phosphorus")),
         Yf = as.factor(Yf),
         streamTreat = as.factor(streamTreat)) %>% 
  filter(treatment == "ambient" | treatment == "phosphorus") %>% 
  droplevels() %>% 
  # to simplify rewriting of code change NEPC to NEP
  mutate(NEP = NEPC)
```

# GPP Plots
Big differences among treatments. Pretty normal
```{r echo=FALSE}
ggplot(met, aes(y = NEP, x = doy, color = treatment)) +
  geom_point() +
  geom_smooth() +
  facet_wrap(vars(stream))
```

```{r echo=FALSE}
ggplot(met, aes(sample = NEP, color = stream)) +stat_qq() + stat_qq_line() +
  facet_grid(.~stream)
```

# GAMMS

```{r echo= FALSE}
BestNEPCor <- corARMA(p = 1,q = 1, form = ~ doy | streamTreat)
```



## Model selection
```{r echo=FALSE, include = T}
g.modsel_t <- uGamm(NEP ~ treatment * invKT.C.StMean  + s(LightPerDay.lcg),
                correlation = BestNEPCor, REML = FALSE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
g.modsel_s <- uGamm(NEP ~ treatment * stream  + s(LightPerDay.lcg),
                correlation = BestNEPCor, REML = FALSE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000), class = "gamm")
# 
# 
g.Dredge_t <-  dredge(global.model = g.modsel_t, evaluate = TRUE, rank = AICc)
g.Dredge_s <-  dredge(global.model = g.modsel_s, evaluate = TRUE, rank = AICc)

```


```{r echo=FALSE}
g.DredgeSumF <- merge(g.Dredge_t, g.Dredge_s)
```




Model selection for all models
```{r echo=FALSE}
subset(g.DredgeSumF, delta <= 2)
```

```{r}
write_csv(g.DredgeSumF,
          file.path(here::here("06_AICtables"), "03a_NEP_P.csv"))
```



```{r}
g.nep.BEST <- gamm(NEP ~ treatment * stream  + s(LightPerDay.lcg),
                correlation = BestNEPCor, REML = TRUE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
```


Stream Temperature models
262.1 v. 284.1
```{r echo=FALSE}
subset(g.Dredge_t, delta <= 2)
```


```{r}
g.nep.BEST_t <- gamm(NEP ~ treatment * invKT.C.StMean    + s(LightPerDay.lcg),
                correlation = BestNEPCor, REML = TRUE, data = met,
                control = lmeControl(maxIter = 1e8, msMaxIter = 1e8, niterEM = 1000))
```

### Best models
#### Overall
```{r}
summary(g.nep.BEST$gam)
```

```{r}
plot(g.nep.BEST$gam,pages =1)
```

Issues with Qres and DOYres
```{r}
round(concurvity(g.nep.BEST$gam, full = FALSE)$estimate,1)
```

```{r}
g.nep.BEST.gv <- getViz(g.nep.BEST$gam)
check(g.nep.BEST.gv)
```



One last check on the autocorrelation. Just foucsing on 3. Nice!

```{r echo= FALSE, include = FALSE}
g.nep.best.acf <- met %>%
  mutate(g.nepRes = resid(g.nep.BEST$lme, type = "normalized")) %>%
  select(streamTreat, g.nepRes) %>%
  group_by(streamTreat) %>%
  nest() %>%
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.nepRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.nep.best.acf$acf, ncol = 3)
```


```{r echo=FALSE}
metBestTest <- met %>% 
  mutate(nepres1 = residuals(g.nep.BEST$gam, type = "response"),
         nepPred1 = predict(g.nep.BEST$gam, type = "response"),
         nepres2 = residuals(g.nep.BEST_t$gam, type = "response"),
         nepPred2 = predict(g.nep.BEST_t$gam, type = "response"))
```


 Check residuals
Still some signal with doy. Also early dates with S18-P giving problems
Not much of a difference between 1 and 2.
```{r echo=FALSE}
ggplot(metBestTest, aes(y = nepres1, x = doy, color = stream, shape = treatment)) +
  geom_point() +
  facet_grid(stream ~ treatment) 
```

```{r echo=FALSE}
ggplot(metBestTest, aes(y = NEP, x = nepPred1, color = stream, shape = treatment)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1)
```


#### stream temp

```{r}
summary(g.nep.BEST_t$gam)
```

Issues with doy and qres
```{r}
round(concurvity(g.nep.BEST_t$gam, full = FALSE)$estimate,1)
```


Some ugliness here
```{r}
g.nep.BEST_t.gv <- getViz(g.nep.BEST_t$gam)
check(g.nep.BEST_t.gv)
```

Nice!
```{r echo= FALSE}
g.nep.BEST_t.acf <- met %>% 
  mutate(g.nepRes = resid(g.nep.BEST_t$lme, type = "normalized")) %>% 
  select(streamTreat, g.nepRes) %>% 
  group_by(streamTreat) %>% 
  nest() %>% 
  mutate(acf = map(.x = data, .f = function(d) ggAcf(d$g.nepRes) + labs(title = streamTreat)))
```

```{r echo= FALSE}
grid.arrange(grobs = g.nep.BEST_t.acf$acf, ncol = 3)
```

```{r echo=FALSE}
ggplot(metBestTest, aes(y = nepres2, x = doy, color = stream, shape = treatment)) +
  geom_point() +
  facet_grid(stream ~ treatment) 
```


```{r echo=FALSE}
ggplot(metBestTest, aes(y = NEP, x = nepPred2, color = stream, shape = treatment)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1)
```




### save results
```{r echo = FALSE, include = FALSE}
 # saveRDS(g.nep.BEST,
 #         file = file.path(here::here("05_SavedModels"), "03a_NEP_P_Best.rds"))
 # saveRDS(g.nep.BEST_t,
 #         file = file.path(here::here("05_SavedModels"), "03a_NEP_P_BestTemp.rds"))
```




```{r}
# save.image(file.path(here::here("04_SavedImages"), "04c_NEPanalysesP_rdat"))
# load(file.path(here::here("04_SavedImages"), "03_GPPanalyses_rdat"))
```
